+++ tests/unit/drivers/prompt_models/test_sagemaker_jumpstart_llama3_instruct_prompt_model_driver.py
@@ -1,16 +1,20 @@
 import boto3
 import pytest
+from griptape.drivers.prompt_model.sagemaker_jumpstart_llama3_instruct_prompt_model_driver import (
+    SageMakerJumpStartLlama3InstructPromptModelDriver,
+)
 from griptape.utils import PromptStack
-from griptape.drivers import AmazonSageMakerPromptDriver, SageMakerFalconPromptModelDriver
+from griptape.drivers import AmazonSageMakerPromptDriver
 
 
-class TestSageMakerFalconPromptModelDriver:
+class TestSageMakerJumpStartLlama3InstructPromptModelDriver:
     @pytest.fixture
     def driver(self):
         return AmazonSageMakerPromptDriver(
             endpoint="endpoint-name",
+            model="inference-component-name",
             session=boto3.Session(region_name="us-east-1"),
-            prompt_model_driver=SageMakerFalconPromptModelDriver(),
+            prompt_model_driver=SageMakerJumpStartLlama3InstructPromptModelDriver(),
             temperature=0.12345,
         ).prompt_model_driver
 
@@ -30,14 +34,19 @@ class TestSageMakerFalconPromptModelDriver:
         model_input = driver.prompt_stack_to_model_input(stack)
 
         assert isinstance(model_input, str)
-        assert model_input.startswith("foo\n\nUser: bar")
+        assert model_input == (
+            "<|begin_of_text|>"
+            "<|start_header_id|>system<|end_header_id|>\n\nfoo<|eot_id|>"
+            "<|start_header_id|>user<|end_header_id|>\n\nbar<|eot_id|>"
+            "<|start_header_id|>assistant<|end_header_id|>\n\n"
+        )
 
     def test_prompt_stack_to_model_params(self, driver, stack):
-        assert driver.prompt_stack_to_model_params(stack)["max_new_tokens"] == 590
+        assert driver.prompt_stack_to_model_params(stack)["max_new_tokens"] == 7991
         assert driver.prompt_stack_to_model_params(stack)["temperature"] == 0.12345
 
     def test_process_output(self, driver, stack):
-        assert driver.process_output([{"generated_text": "foobar"}]).value == "foobar"
+        assert driver.process_output({"generated_text": "foobar"}).value == "foobar"
 
     def test_tokenizer_max_model_length(self, driver):
-        assert driver.tokenizer.tokenizer.model_max_length == 2048
+        assert driver.tokenizer.tokenizer.model_max_length == 8000