+++ tests/unit/drivers/prompt_models/test_sagemaker_llama_prompt_model_driver.py
@@ -8,7 +8,7 @@ class TestSageMakerLlamaPromptModelDriver:
     @pytest.fixture
     def driver(self):
         return AmazonSageMakerPromptDriver(
-            model="foo",
+            endpoint="endpoint-name",
             session=boto3.Session(region_name="us-east-1"),
             prompt_model_driver=SageMakerLlamaPromptModelDriver(),
             temperature=0.12345,